requirements.txt
spotipy==2.22.1
requests==2.31.0
twitchAPI==4.1.0
configparser==6.0.0
pyi18n-v2==1.2.1
apscheduler==3.10.4
PyInstaller==6.3.0